gamesearch/requirements.txt
django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.8
python-dotenv==1.0.0
django-cors-headers==3.14.0
